pyproject.toml
[project]
name = "fastapi-ca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "jq>=1.10.0",
    "passlib[bcrypt]>=1.7.4",
    "py-ulid>=1.0.3",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
